# Remove `NewPossibleNonCounterInfo` and minimise creating empty annotations