Follow up to https://github.com/prometheus/prometheus/pull/12959 as a quick fix to temporarily remove `NewPossibleNonCounterInfo` and avoid creating empty annotations unnecessarily.

Followed up with https://github.com/prometheus/prometheus/pull/13022 to put `NewPossibleNonCounterInfo` back without running it every step, and apply a similar treatment for other warnings which don't look at the data like `NewInvalidQuantileWarning`.

According to benchmarks, this should restore most of the efficiency:
Before is run on [a branch on main that reverts the commits that added and modified annotations](https://github.com/zenador/prometheus/tree/temp-benchmark-remove-annotations), and after is run on [a branch on main with this commit applied](https://github.com/zenador/prometheus/tree/remove-NewPossibleNonCounterInfo).
```
goos: linux
goarch: amd64
pkg: github.com/prometheus/prometheus/promql
cpu: 11th Gen Intel(R) Core(TM) i7-11800H @ 2.30GHz
                                                  │ BenchmarkRangeQuery_before.txt │   BenchmarkRangeQuery_after.txt    │
                                                  │             sec/op             │   sec/op     vs base               │
RangeQuery/expr=rate(a_hundred[1m]),steps=1000-16                      9.176m ± 4%   9.645m ± 1%  +5.11% (p=0.002 n=10)

                                                  │ BenchmarkRangeQuery_before.txt │ BenchmarkRangeQuery_after.txt  │
                                                  │              B/op              │     B/op      vs base          │
RangeQuery/expr=rate(a_hundred[1m]),steps=1000-16                     367.6Ki ± 0%   367.6Ki ± 0%  ~ (p=0.353 n=10)

                                                  │ BenchmarkRangeQuery_before.txt │   BenchmarkRangeQuery_after.txt    │
                                                  │           allocs/op            │  allocs/op   vs base               │
RangeQuery/expr=rate(a_hundred[1m]),steps=1000-16                      5.256k ± 0%   5.257k ± 0%  +0.02% (p=0.000 n=10)
```